create src.exceptions module, move modal exceptions there
