+++ garden-backend-service/src/middleware/logging.py
@@ -1,6 +1,4 @@
-import re
 import time
-import traceback
 import uuid
 
 import structlog
@@ -8,35 +6,8 @@ from fastapi import Request, Response
 from fastapi.responses import JSONResponse
 from starlette.middleware.base import BaseHTTPMiddleware
 
-from src.api.dependencies.modal import ModalException, handle_modal_exception
-
-
-def filter_stacktrace(exc: Exception, func_to_stop_at: str = "run_endpoint_function"):
-    """Filter the stack trace to include only the relevant part from the endpoint call to the error."""
-    tb = traceback.extract_tb(exc.__traceback__)  # Extract the stack frames
-
-    # Find the frame for the most recent endpoint call
-    relevant_frames = []
-
-    for frame in reversed(tb):  # Walk the stack from the bottom (most recent call)
-        filename, lineno, funcname, text = frame
-
-        # stop when we hit the inital call to the endpoint
-        if funcname == func_to_stop_at:
-            break
-
-        relevant_frames.append(frame)
-
-    # If we found relevant frames, remove the noisy stuff
-    if relevant_frames:
-        filtered_tb = traceback.format_list(
-            reversed(relevant_frames)
-        )  # Reverse to maintain original order
-        cleaned_tb = [re.sub(r"(\^+)", "", line).strip() for line in filtered_tb]
-        return "".join(cleaned_tb)
-
-    # Fallback to the full traceback if no endpoint call was found
-    return "".join(traceback.format_exception(type(exc), exc, exc.__traceback__))
+from src.exceptions.modal import ModalException, handle_modal_exception
+from src.exceptions.utils import format_traceback
 
 
 class LogRequestIdMiddleware(BaseHTTPMiddleware):
@@ -93,7 +64,7 @@ class ErrorHandlingMiddleware(BaseHTTPMiddleware):
             return await call_next(request)
         except ModalException as e:
             logger = structlog.get_logger()
-            filtered_stacktrace = filter_stacktrace(e)
+            filtered_stacktrace = format_traceback(e)
             logger.error(
                 "Modal Exception",
                 method=request.method,
@@ -105,7 +76,7 @@ class ErrorHandlingMiddleware(BaseHTTPMiddleware):
             )
             return handle_modal_exception(request, e)
         except Exception as e:
-            filtered_stacktrace = filter_stacktrace(e)
+            filtered_stacktrace = format_traceback(e)
             logger = structlog.get_logger()
             logger.error(
                 "Unhandled exception",
